feat: 문제 풀이 화면 레이아웃

+++ lib/features/question/util/basic_operation_util.dart
@@ -22,13 +22,13 @@ int getDigitNumber(DigitCount digitCount) {
 double calculator(num1, num2, operator) {
   switch (operator) {
     case ArithmeticType.addition:
-      return num1 + num2;
+      return (num1 + num2).toDouble();
     case ArithmeticType.subtraction:
-      return num1 - num2;
+      return num1 - num2.toDouble();
     case ArithmeticType.multiplication:
-      return num1 * num2;
+      return num1 * num2.toDouble();
     case ArithmeticType.division:
-      return num1 / num2;
+      return num1 / num2.toDouble();
   }
   return -1;
 }
